fix(review): consolidate shared fixtures to conftest, remove test slop
- Move skill_text and report_section fixtures to tests/skills/conftest.py
  to eliminate verbatim duplication across test_investigate_contracts.py and
  test_investigate_deep_mode_contracts.py
- Remove test_deep_workflow_section_exists (redundant: deep_workflow_section
  fixture already fails if the heading is absent and all D1-D6 tests depend on it)
- Replace six structurally identical D1-D6 one-liners with a single
  @pytest.mark.parametrize test
- Tighten has_prohibition check to 'automatically' only, removing the
  trivially-True 'never'/'not' conditions
